[feat] Basic event views complete [feat] Separated event views into different components [feat] Connected trades to store and server
Note:

Still need to implement audit view. Will need to discuss how we want to handle this.
