+++ client/src/App.vue
@@ -70,7 +70,7 @@ body {
   --marslog-purple: rgba(111, 36, 134, 0.05);
 
   --font-family-default: 'Space Mono', monospace;
-  --font-small: 0.86rem;
+  --font-small: 0.75rem;
   --font-med: 1rem;
   --font-large: 1.5rem;
