+++ shared/types.ts
@@ -87,7 +87,7 @@ export enum EventServerActions {
 
 export enum EventClientView {
   NO_CHANGE = 'NO_CHANGE',
-  YES_NO_VOTE = 'YES_NO_VOTE',
+  VOTE_YES_NO = 'VOTE_YES_NO',
   SELECT_INFLUENCES = 'SELECT_INFLUENCES',
   DRAW_INFLUENCES = 'DRAW_INFLUENCES',
   VOTE_FOR_PLAYER_SINGLE = 'VOTE_FOR_PLAYER_SINGLE',
@@ -95,7 +95,6 @@ export enum EventClientView {
   AUDIT = 'AUDIT',
   SELECT_PURCHASED_ACCOMPLISHMENT = 'SELECT_PURCHASED_ACCOMPLISHMENT',
   DISABLE_CHAT = 'DISABLE_CHAT',
-  EVENT_TEST = 'EVENT_TEST'
 }
 
 export enum EventClientActions {
